[Improvement-17330][K8s] Extract job monitor for k8s task
