+++ dolphinscheduler-task-plugin/dolphinscheduler-task-api/src/test/java/org/apache/dolphinscheduler/plugin/task/api/k8s/K8sTaskExecutorTest.java
@@ -17,32 +17,14 @@
 
 package org.apache.dolphinscheduler.plugin.task.api.k8s;
 
-import static org.apache.dolphinscheduler.plugin.task.api.TaskConstants.EXIT_CODE_FAILURE;
-import static org.apache.dolphinscheduler.plugin.task.api.TaskConstants.EXIT_CODE_SUCCESS;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.eq;
-import static org.mockito.Mockito.doThrow;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
-
 import org.apache.dolphinscheduler.plugin.task.api.TaskException;
 import org.apache.dolphinscheduler.plugin.task.api.TaskExecutionContext;
-import org.apache.dolphinscheduler.plugin.task.api.enums.TaskTimeoutStrategy;
 import org.apache.dolphinscheduler.plugin.task.api.k8s.impl.K8sTaskExecutor;
-import org.apache.dolphinscheduler.plugin.task.api.model.TaskResponse;
-import org.apache.dolphinscheduler.plugin.task.api.utils.K8sUtils;
 
-import java.lang.reflect.Field;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.TimeUnit;
-import java.util.concurrent.atomic.AtomicReference;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
@@ -50,9 +32,6 @@ import org.junit.jupiter.api.Test;
 
 import io.fabric8.kubernetes.api.model.NodeSelectorRequirement;
 import io.fabric8.kubernetes.api.model.batch.v1.Job;
-import io.fabric8.kubernetes.api.model.batch.v1.JobStatus;
-import io.fabric8.kubernetes.client.informers.ResourceEventHandler;
-import io.fabric8.kubernetes.client.informers.SharedIndexInformer;
 
 public class K8sTaskExecutorTest {
 
@@ -66,7 +45,6 @@ public class K8sTaskExecutorTest {
     private final int taskInstanceId = 1000;
     private final String taskName = "k8s_task_test";
     private Job job;
-    private K8sUtils k8sUtils;
 
     @BeforeEach
     public void before() throws Exception {
@@ -81,8 +59,6 @@ public class K8sTaskExecutorTest {
         requirement.setOperator("In");
         requirement.setValues(Arrays.asList("1234", "123456"));
         k8sTaskExecutor = new K8sTaskExecutor(taskRequest);
-        k8sUtils = mock(K8sUtils.class);
-        injectK8sUtils(k8sTaskExecutor, k8sUtils);
         k8sTaskMainParameters = new K8sTaskMainParameters();
         k8sTaskMainParameters.setImage(image);
         k8sTaskMainParameters.setImagePullPolicy(imagePullPolicy);
@@ -96,246 +72,6 @@ public class K8sTaskExecutorTest {
         job = k8sTaskExecutor.getJob();
     }
 
-    private void injectK8sUtils(K8sTaskExecutor executor, K8sUtils mockK8sUtils) throws Exception {
-        Class<?> type = executor.getClass();
-        while (type != null) {
-            try {
-                Field field = type.getDeclaredField("k8sUtils");
-                field.setAccessible(true);
-                field.set(executor, mockK8sUtils);
-                return;
-            } catch (NoSuchFieldException ignored) {
-                type = type.getSuperclass();
-            }
-        }
-        throw new NoSuchFieldException("k8sUtils");
-    }
-
-    private TaskExecutionContext getTaskRequest() throws Exception {
-        Field field = k8sTaskExecutor.getClass().getSuperclass().getDeclaredField("taskRequest");
-        field.setAccessible(true);
-        return (TaskExecutionContext) field.get(k8sTaskExecutor);
-    }
-
-    private WatcherHarness startBatchJobWatcher(TaskResponse taskResponse) throws InterruptedException {
-        return startBatchJobWatcher(k8sTaskExecutor, taskResponse);
-    }
-
-    private WatcherHarness startBatchJobWatcher(K8sTaskExecutor executor,
-                                                TaskResponse taskResponse) throws InterruptedException {
-        WatcherHarness harness = new WatcherHarness();
-        harness.informer = mock(SharedIndexInformer.class);
-        when(harness.informer.start()).thenReturn(CompletableFuture.completedFuture(null));
-        CountDownLatch handlerReady = new CountDownLatch(1);
-        AtomicReference<ResourceEventHandler<Job>> handlerRef = new AtomicReference<>();
-        when(k8sUtils.createBatchJobInformer(eq(job.getMetadata().getName()), eq(namespace), any()))
-                .thenAnswer(invocation -> {
-                    handlerRef.set(invocation.getArgument(2));
-                    handlerReady.countDown();
-                    return harness.informer;
-                });
-        harness.thread = new Thread(() -> executor.registerBatchJobWatcher(job, taskResponse));
-        harness.thread.start();
-        Assertions.assertTrue(handlerReady.await(5, TimeUnit.SECONDS));
-        harness.handler = handlerRef.get();
-        return harness;
-    }
-
-    private void finishWatcher(WatcherHarness harness) throws InterruptedException {
-        harness.thread.join(5000);
-        verify(harness.informer).stop();
-    }
-
-    private Job jobWithStatus(Integer succeeded, Integer failed) {
-        JobStatus status = new JobStatus();
-        status.setSucceeded(succeeded);
-        status.setFailed(failed);
-        Job watchedJob = new Job();
-        watchedJob.setMetadata(job.getMetadata());
-        watchedJob.setStatus(status);
-        return watchedJob;
-    }
-
-    @Test
-    public void testRegisterBatchJobInformerOnUpdateSuccess() throws Exception {
-        TaskResponse taskResponse = new TaskResponse();
-        WatcherHarness harness = startBatchJobWatcher(taskResponse);
-        harness.handler.onUpdate(job, jobWithStatus(1, null));
-        finishWatcher(harness);
-        assertEquals(EXIT_CODE_SUCCESS, taskResponse.getExitStatusCode());
-    }
-
-    @Test
-    public void testRegisterBatchJobInformerOnUpdateFailed() throws Exception {
-        TaskResponse taskResponse = new TaskResponse();
-        WatcherHarness harness = startBatchJobWatcher(taskResponse);
-        harness.handler.onUpdate(job, jobWithStatus(null, 1));
-        finishWatcher(harness);
-        assertEquals(EXIT_CODE_FAILURE, taskResponse.getExitStatusCode());
-    }
-
-    @Test
-    public void testRegisterBatchJobInformerOnDelete() throws Exception {
-        TaskResponse taskResponse = new TaskResponse();
-        WatcherHarness harness = startBatchJobWatcher(taskResponse);
-        harness.handler.onDelete(job, false);
-        finishWatcher(harness);
-        assertEquals(EXIT_CODE_FAILURE, taskResponse.getExitStatusCode());
-    }
-
-    @Test
-    public void testRegisterBatchJobInformerOnAddSuccess() throws Exception {
-        TaskResponse taskResponse = new TaskResponse();
-        WatcherHarness harness = startBatchJobWatcher(taskResponse);
-        harness.handler.onAdd(jobWithStatus(1, null));
-        finishWatcher(harness);
-        assertEquals(EXIT_CODE_SUCCESS, taskResponse.getExitStatusCode());
-    }
-
-    @Test
-    public void testRegisterBatchJobInformerOnAddFailed() throws Exception {
-        TaskResponse taskResponse = new TaskResponse();
-        WatcherHarness harness = startBatchJobWatcher(taskResponse);
-        harness.handler.onAdd(jobWithStatus(null, 1));
-        finishWatcher(harness);
-        assertEquals(EXIT_CODE_FAILURE, taskResponse.getExitStatusCode());
-    }
-
-    @Test
-    public void testRegisterBatchJobInformerOnAddRunning() throws Exception {
-        TaskResponse taskResponse = new TaskResponse();
-        WatcherHarness harness = startBatchJobWatcher(taskResponse);
-        harness.handler.onAdd(jobWithStatus(null, null));
-        Assertions.assertTrue(harness.thread.isAlive());
-        harness.handler.onUpdate(job, jobWithStatus(1, null));
-        finishWatcher(harness);
-        assertEquals(EXIT_CODE_SUCCESS, taskResponse.getExitStatusCode());
-    }
-
-    @Test
-    public void testRegisterBatchJobInformerInformerStartFailed() throws Exception {
-        TaskResponse taskResponse = new TaskResponse();
-        doThrow(new RuntimeException("informer start failed")).when(k8sUtils)
-                .createBatchJobInformer(eq(job.getMetadata().getName()), eq(namespace), any());
-        Thread thread = new Thread(() -> k8sTaskExecutor.registerBatchJobWatcher(job, taskResponse));
-        thread.start();
-        thread.join(5000);
-        assertEquals(EXIT_CODE_FAILURE, taskResponse.getExitStatusCode());
-    }
-
-    @Test
-    public void testRegisterBatchJobInformerIgnoreRunningUpdate() throws Exception {
-        TaskResponse taskResponse = new TaskResponse();
-        WatcherHarness harness = startBatchJobWatcher(taskResponse);
-        harness.handler.onUpdate(job, jobWithStatus(null, null));
-        Assertions.assertTrue(harness.thread.isAlive());
-        harness.handler.onUpdate(job, jobWithStatus(1, null));
-        finishWatcher(harness);
-        assertEquals(EXIT_CODE_SUCCESS, taskResponse.getExitStatusCode());
-    }
-
-    @Test
-    public void testRegisterBatchJobInformerTimeout() throws Exception {
-        TaskExecutionContext taskRequest = getTaskRequest();
-        taskRequest.setTaskTimeoutStrategy(TaskTimeoutStrategy.FAILED);
-        taskRequest.setTaskTimeout(1);
-
-        TaskResponse taskResponse = new TaskResponse();
-        WatcherHarness harness = startBatchJobWatcher(taskResponse);
-        finishWatcher(harness);
-        assertEquals(EXIT_CODE_FAILURE, taskResponse.getExitStatusCode());
-    }
-
-    @Test
-    public void testRegisterBatchJobInformerJobStatusPollingSuccess() throws Exception {
-        K8sTaskExecutor pollingExecutor = new K8sTaskExecutor(getTaskRequest()) {
-
-            @Override
-            protected long getJobStatusPollIntervalSeconds() {
-                return 1L;
-            }
-        };
-        injectK8sUtils(pollingExecutor, k8sUtils);
-        when(k8sUtils.getJob(eq(job.getMetadata().getName()), eq(namespace)))
-                .thenReturn(jobWithStatus(1, null));
-
-        TaskResponse taskResponse = new TaskResponse();
-        WatcherHarness harness = startBatchJobWatcher(pollingExecutor, taskResponse);
-        finishWatcher(harness);
-        assertEquals(EXIT_CODE_SUCCESS, taskResponse.getExitStatusCode());
-        verify(k8sUtils).getJob(eq(job.getMetadata().getName()), eq(namespace));
-    }
-
-    @Test
-    public void testRegisterBatchJobInformerJobStatusPollingDeleted() throws Exception {
-        K8sTaskExecutor pollingExecutor = new K8sTaskExecutor(getTaskRequest()) {
-
-            @Override
-            protected long getJobStatusPollIntervalSeconds() {
-                return 1L;
-            }
-        };
-        injectK8sUtils(pollingExecutor, k8sUtils);
-        when(k8sUtils.getJob(eq(job.getMetadata().getName()), eq(namespace))).thenReturn(null);
-
-        TaskResponse taskResponse = new TaskResponse();
-        WatcherHarness harness = startBatchJobWatcher(pollingExecutor, taskResponse);
-        finishWatcher(harness);
-        assertEquals(EXIT_CODE_FAILURE, taskResponse.getExitStatusCode());
-        verify(k8sUtils).getJob(eq(job.getMetadata().getName()), eq(namespace));
-    }
-
-    @Test
-    public void testRegisterBatchJobInformerConsecutivePollFailuresWithoutTimeoutStrategy() throws Exception {
-        TaskExecutionContext taskRequest = getTaskRequest();
-        taskRequest.setTaskTimeoutStrategy(TaskTimeoutStrategy.WARN);
-
-        K8sTaskExecutor pollingExecutor = new K8sTaskExecutor(taskRequest) {
-
-            @Override
-            protected long getJobStatusPollIntervalSeconds() {
-                return 1L;
-            }
-
-            @Override
-            protected int getMaxConsecutivePollFailures() {
-                return 2;
-            }
-        };
-        injectK8sUtils(pollingExecutor, k8sUtils);
-        when(k8sUtils.getJob(eq(job.getMetadata().getName()), eq(namespace)))
-                .thenThrow(new RuntimeException("apiserver unreachable"));
-
-        TaskResponse taskResponse = new TaskResponse();
-        WatcherHarness harness = startBatchJobWatcher(pollingExecutor, taskResponse);
-        finishWatcher(harness);
-        assertEquals(EXIT_CODE_FAILURE, taskResponse.getExitStatusCode());
-        verify(k8sUtils, org.mockito.Mockito.atLeast(2))
-                .getJob(eq(job.getMetadata().getName()), eq(namespace));
-    }
-
-    private static final class WatcherHarness {
-
-        private SharedIndexInformer<Job> informer;
-        private ResourceEventHandler<Job> handler;
-        private Thread thread;
-    }
-
-    @Test
-    public void testGetK8sJobStatusNormal() {
-        JobStatus jobStatus = new JobStatus();
-        jobStatus.setSucceeded(1);
-        job.setStatus(jobStatus);
-        Assertions.assertEquals(0, Integer.compare(0, k8sTaskExecutor.getK8sJobStatus(job)));
-    }
-    @Test
-    public void testSetTaskStatusNormal() {
-        int jobStatus = 0;
-        TaskResponse taskResponse = new TaskResponse();
-        k8sTaskExecutor.setJob(job);
-        k8sTaskExecutor.setTaskStatus(jobStatus, job.getMetadata().getName(), taskResponse);
-        Assertions.assertEquals(0, taskResponse.getExitStatusCode());
-    }
     @Test
     public void testWaitTimeoutNormal() {
         try {
@@ -352,5 +88,4 @@ public class K8sTaskExecutorTest {
                 k8sTaskExecutor.getJob().getSpec().getTemplate().getSpec().getContainers().get(0).getCommand();
         Assertions.assertEquals(expectedCommands, actualCommands);
     }
-
 }
